Create Transform TO left- cloned.java
In this solution we are making clone of left node's data. for this we have used a trick in this trick we have created a new node on line number 91 in this node data will be same as previous node and in line 93 we have joined previous node with the new node and we have created left node's copy.